requirements.txt
pandas
python-dotenv
openai
autogen-agentchat
autogen-ext
playwright
